Move YesNoPrompt to extension method
